Fix tests

+++ onnx/backend/test/case/node/bernoulli.py
@@ -25,7 +25,7 @@ class Bernoulli(Base):
             outputs=["y"],
         )
 
-        x = np.random.uniform(0.0, 1.0, 10).astype(np.float32)
+        x = np.random.uniform(0.0, 1.0, 10).astype(float)
         y = bernoulli_reference_implementation(x, float)
         expect(node, inputs=[x], outputs=[y], name="test_bernoulli")
 
@@ -53,5 +53,5 @@ class Bernoulli(Base):
         )
 
         x = np.random.uniform(0.0, 1.0, 10).astype(np.float32)
-        y = bernoulli_reference_implementation(x, float)
+        y = bernoulli_reference_implementation(x, np.float32)
         expect(node, inputs=[x], outputs=[y], name="test_bernoulli_seed")
